requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# AWS SDK
boto3>=1.28.0

# Audio processing
faster-whisper>=0.10.0

# Video/Image processing
opencv-python>=4.8.0
imagehash>=4.3.1
Pillow>=10.0.0

# File watching
watchdog>=3.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
